feat(windows-fullscreen-detector): update to version 26.0.1 and add fullscreen_core.c
- Bumped the version of @eisland/windows-fullscreen-detector to 26.0.1 in package.json and package-lock.json.
- Added a new source file, fullscreen_core.c, to enhance fullscreen detection functionality.
- Updated binding.gyp to include fullscreen_core.c as a source file for the module.
