+++ plugins/windows-fullscreen-detector/binding.gyp
@@ -2,7 +2,7 @@
   "targets": [
     {
       "target_name": "windows_fullscreen_detector",
-      "sources": ["src/fullscreen_detector.c"],
+      "sources": ["src/fullscreen_detector.c", "src/fullscreen_core.c"],
       "defines": ["WIN32_LEAN_AND_MEAN", "NAPI_VERSION=8"],
       "libraries": ["user32.lib", "dwmapi.lib"]
     }
