+++ plugins/windows-fullscreen-detector/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eisland/windows-fullscreen-detector",
-  "version": "26.0.0",
+  "version": "26.0.1",
   "description": "Windows N-API addon for detecting foreground fullscreen windows across multiple monitors.",
   "author": "JNTMTMTM",
   "license": "GPL-3.0",
